// client/src/model/dragDimension.ts
export default class dragDimension
{
  initY: number = 0; // Initial Drag Selection Y coordinate
  initX: number = 0; // Initial Drag Selection X coordinate
  top: number = 0; // Drag selector top
  left: number = 0; // Drag selector left
  width: number = 0; // Drag selector width
  height: number = 0; // Drag selector height
}